first working version? (Without formal tests)
