+++ src/frompackage/helpers.jl
@@ -31,18 +31,6 @@ function simulate_manual_rerun(cell_ids::Array; kwargs...)
 end
 =#
 
-# Functions to add and remove from the LOAD_PATH
-function add_loadpath(entry::String; should_prepend)
-    idx = findfirst(==(entry), LOAD_PATH)
-    if isnothing(idx)
-        f! = should_prepend ? pushfirst! : push!
-        # We add
-        f!(LOAD_PATH, entry)
-    end
-    return nothing
-end
-add_loadpath(ecg::EnvCacheGroup; kwargs...) = add_loadpath(ecg |> get_active |> get_project_file; kwargs...)
-
 ## execute only in notebook
 # We have to create our own simple check to only execute some stuff inside the notebook where they are defined. We have stuff in basics.jl but we don't want to include that in this notebook
 function is_notebook_local(calling_file::String)
@@ -50,14 +38,14 @@ function is_notebook_local(calling_file::String)
     return length(name_cell) == 2 && length(name_cell[2]) == 36
 end
 
-# Get the first element in itr that satisfies predicate p, or nothing if itr is empty or no elements satisfy p
-function getfirst(p, itr)
-    for el in itr
-        p(el) && return el
-    end
-    return nothing
-end
-getfirst(itr) = getfirst(x -> true, itr)
+# # Get the first element in itr that satisfies predicate p, or nothing if itr is empty or no elements satisfy p
+# function getfirst(p, itr)
+#     for el in itr
+#         p(el) && return el
+#     end
+#     return nothing
+# end
+# getfirst(itr) = getfirst(x -> true, itr)
 
 
 ## HTML Popup
@@ -139,9 +127,6 @@ function issamepath(path1::String, path2::String)
 end
 issamepath(path1::Symbol, path2::Symbol) = issamepath(String(path1), String(path2))
 
-# Create a Base.PkgId from a PkgInfo
-to_pkgid(p::PkgInfo) = Base.PkgId(p.uuid, p.name)
-
 # This will extract the string from a raw_str macro, and will throw an error otherwise
 function extract_raw_str(ex::Expr)
     valid = Meta.isexpr(ex, :macrocall) && ex.args[1] === Symbol("@raw_str")
@@ -153,34 +138,6 @@ function extract_raw_str(ex::Expr)
 end
 extract_raw_str(s::AbstractString) = String(s), true
 
-function get_extensions_ids(old_module::Module, parent::Base.PkgId)
-    package_dict = old_module._fromparent_dict_
-    out = Base.PkgId[]
-    if has_extensions(package_dict)
-        for ext in keys(package_dict["extensions"])
-            id = Base.PkgId(Base.uuid5(parent.uuid, ext), ext)
-            push!(out, id)
-        end
-    end
-    return out
-end
-
-# This function will get the module stored in the created_modules dict based on the entry point
-get_stored_module() = STORED_MODULE[]
-# This will store in it
-update_stored_module(m::Module) = STORED_MODULE[] = m
-function update_stored_module(package_dict::Dict)
-    m = get_target_module(package_dict)
-    update_stored_module(m)
-end
-
-overwrite_imported_symbols(package_dict::Dict) = overwrite_imported_symbols(get(Set{Symbol}, package_dict, "Catchall Imported Symbols"))
-# This overwrites the PREVIOUSLY_IMPORTED_SYMBOLS with the contents of new_symbols
-function overwrite_imported_symbols(new_symbols)
-    empty!(PREVIOUS_CATCHALL_NAMES)
-    union!(PREVIOUS_CATCHALL_NAMES, new_symbols)
-    nothing
-end
 
 function beautify_package_path(p::FromPackageController{name}) where name
     @nospecialize
@@ -224,4 +181,147 @@ notebookObserver.observe(notebook, {childList: true})
     	replaceTextInNode(content, regex, replacement);
     </script>
 """
+end
+
+function generate_manifest_deps(proj_file::String)
+    envdir = dirname(abspath(proj_file))
+    manifest_file = ""
+    for name in ("Manifest.toml", "JuliaManifest.toml")
+        path = joinpath(envdir, name)
+        if isfile(path)
+            manifest_file = path
+            break
+        end
+    end
+    @assert !isempty(manifest_file) "A manifest could not be found at the project's location.\nYou have to provide an instantiated environment."
+    d = TOML.parsefile(manifest_file)
+    out = Dict{Base.UUID, String}()
+    for (name, data) in d["deps"]
+        # We use only here because I believe the entry will always contain a single dict wrapped in an array. If we encounter a case where this is not true the only will throw instead of silently taking just the first
+        uuid = only(data)["uuid"] |> Base.UUID
+        out[uuid] = name
+    end
+    return out
+end
+
+function update_loadpath(p::FromPackageController)
+    @nospecialize
+    proj_file = p.project.file 
+    if proj_file ∉ LOAD_PATH
+        push!(LOAD_PATH, proj_file)
+    end
+end
+
+nested_getproperty_expr(name::Symbol) = QuoteNode(name)
+# This function creates the expression to access a nested property specified by a path. For example, if `path = [:Main, :ASD, :LOL]`, `nested_getproperty_expr(path...)` will return the expression equivalent to `Main.ASD.LOL`. This is not to be used within `import/using` statements as the synthax for accessing nested modules is different there.
+function nested_getproperty_expr(names_path::Symbol...)
+    @nospecialize
+    others..., tail = names_path
+    last_arg = nested_getproperty_expr(tail)
+    first_arg = length(others) === 1 ? first(others) : nested_getproperty_expr(others...)
+    ex = isempty(others) ? arg : Expr(:., first_arg, last_arg)
+    return ex
+end
+
+### Input Parsing
+# This function will extract the first name of a module identifier from `import/using` statements
+function get_modpath_root(ex::Expr)
+    (;modname_path) = extract_input_import_names(ex)
+    modname_first = first(modname_path)
+    return modname_first
+end
+
+# This function traverse a path to access a nested module from a `starting_module`. It is used to extract the corresponding module from `import/using` statements.
+function extract_nested_module(starting_module::Module, nested_path; first_dot_skipped=false)
+    m = starting_module
+    for name in nested_path
+        m = if name === :.
+            first_dot_skipped ? parentmodule(m) : m
+        else
+            @assert isdefined(m, name) "The module `$name` could not be found inside parent module `$(nameof(m))`"
+            getproperty(m, name)::Module
+        end
+        first_dot_skipped = true
+    end
+    return m
+end
+
+function get_temp_module()
+    if isdefined(Main, TEMP_MODULE_NAME)
+        getproperty(Main, TEMP_MODULE_NAME)::Module
+    else
+        m = Core.eval(Main, :(module $TEMP_MODULE_NAME
+        module _LoadedModules_ end
+        end))::Module
+    end
+end
+get_temp_module(s::Symbol) = getproperty(get_temp_module(), s)
+function get_temp_module(names::Vector{Symbol})
+    out = get_temp_module()
+    for name in names
+        getproperty(out, name)
+    end
+    return out
+end
+function get_temp_module(::FromPackageController{name}) where {name}
+    @nospecialize
+    get_temp_module(name)::Module
+end
+
+get_loaded_modules_mod() = get_temp_module(:_LoadedModules_)::Module
+
+
+function populate_loaded_modules()
+    loaded_modules = get_loaded_modules_mod()
+    @lock Base.require_lock begin
+        for (id, m) in Base.loaded_modules
+            name = Symbol(id)
+            isdefined(loaded_modules, name) && continue
+            Core.eval(loaded_modules, :(const $name = $m))
+        end
+    end
+    empty!(Base.package_callbacks) ### IMPORTANT, TO REMOVE ###
+    if mirror_package_callback ∉ Base.package_callbacks
+        # Add the package callback if not already present
+        push!(Base.package_callbacks, mirror_package_callback)
+    end
+end
+
+function get_dep_from_manifest(p::FromPackageController, base_name)
+    @nospecialize
+    (; manifest_deps) = p
+    name_str = string(base_name)
+    for (uuid, name) in manifest_deps
+        if name === name_str
+            id = Base.PkgId(uuid, name)
+            return get_dep_from_loaded_modules(id)
+        end
+    end
+    return nothing
+end
+function get_dep_from_loaded_modules(id::Base.PkgId)
+    loaded_modules = get_loaded_modules_mod()
+    key = Symbol(id)
+    isdefined(loaded_modules, key) || error("The module $key can not be found in the loaded modules.")
+    m = getproperty(loaded_modules, Symbol(id))::Module
+    return m
+end
+function get_dep_from_loaded_modules(p::FromPackageController{name}, base_name; allow_manifest=false, allow_stdlibs=true)::Module where {name}
+    @nospecialize
+    base_name === name && return get_temp_module(p)
+    package_name = string(base_name)
+    if allow_stdlibs
+        uuid = get(STDLIBS_DATA, package_name, nothing)
+        uuid !== nothing && return get_dep_from_loaded_modules(Base.PkgId(uuid, package_name))
+    end
+    proj = p.project
+    uuid = get(proj.deps, package_name) do
+        get(proj.weakdeps, package_name) do
+            out = allow_manifest ? get_dep_from_manifest(p, base_name) : nothing
+            isnothing(out) && error("The package with name $package_name could not be found as deps or weakdeps of the target project, as indirect dep of the manifest, or as standard library")
+            return out
+        end
+    end
+    id = Base.PkgId(uuid, package_name)
+    return get_dep_from_loaded_modules(id)
 end
